Text position and color
I thought this might be a better position and color if you like it. The old positioning I think was a little strange because it was right where the player moves his hand and items which makes it harder to read, and the color I thought was maybe just unnecessary. I tried real quick to get the darkness on the white text right to minimize the color blending of the background and the text.